+++ src/main/gen/com/github/nthykier/debpkg/dch/psi/DchChangelogEntry.java
@@ -4,8 +4,10 @@ package com.github.nthykier.debpkg.dch.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.NavigatablePsiElement;
+import com.intellij.navigation.ItemPresentation;
 
-public interface DchChangelogEntry extends PsiElement {
+public interface DchChangelogEntry extends NavigatablePsiElement {
 
   @NotNull
   List<DchChangelogLine> getChangelogLineList();
@@ -15,5 +17,13 @@ public interface DchChangelogEntry extends PsiElement {
 
   @NotNull
   DchVersionLine getVersionLine();
+
+  boolean isFirstChangelogEntry();
+
+  boolean isUnreleasedEntry();
+
+  ItemPresentation getPresentation();
+
+  DchChangelogEntry getFirstEntry();
 
 }