+++ src/main/gen/com/github/nthykier/debpkg/dch/psi/DchVersion.java
@@ -4,10 +4,7 @@ package com.github.nthykier.debpkg.dch.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiReference;
 
-public interface DchSource extends PsiElement {
-
-  @Nullable PsiReference getReference();
+public interface DchVersion extends PsiElement {
 
 }