+++ src/main/gen/com/github/nthykier/debpkg/dch/psi/impl/DchVersionLineImpl.java
@@ -39,4 +39,10 @@ public class DchVersionLineImpl extends ASTWrapperPsiElement implements DchVersi
     return findNotNullChildByClass(DchSource.class);
   }
 
+  @Override
+  @Nullable
+  public DchVersion getVersion() {
+    return findChildByClass(DchVersion.class);
+  }
+
 }