+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 # debpkg Changelog
 
 ## [Unreleased]
+### Added
+- Structure view for changelog files to provide an overview of the versions.
+
+### Changed
 - Bump the compat range for IDE to support 2022.1
 - Internal/Chore: Update build time dependency on `gradle-intellij-plugin`.
 - Internal/Chore: Update the CI dependency on `actions/checkout`, `actions/cache`, `actions/setup-java`, and `actions/upload-artifact`.